Crate fj_viewer

Fornjot Model Viewer

Fornjot is an early-stage b-rep CAD kernel written in Rust. The kernel is split into multiple libraries that can be used semi-independently, and this is one of those.

This library provides a viewer for Fornjot models.

Structs

ViewerHandle

Handle to the model viewer

WindowHandle

Handle to a model viewer window

Enums

Error

Main loop initialization error

RendererInitError

Error describing the set of render surface initialization errors

Functions

make_viewer_and_spawn_thread

Create a model viewer and spawn a new thread where to use it