fj_viewer

Function make_viewer_and_spawn_thread

pub fn make_viewer_and_spawn_thread<R>(
    f: impl FnOnce(ViewerHandle) -> R + Send + 'static,
) -> Result<R, Error>
where
    R: Send + 'static,

Create a model viewer and spawn a new thread where to use it

Create a model viewer that runs on the main thread, blocking the thread that calls this function. Makes a handle to the viewer available to the provided closure, allowing it to display models.

This API is a bit weird, due to the realities of native graphics on various platforms. Those tend to insist on running on the main thread, so this function spawns a new thread for the caller.

This function should be called from the application’s main thread, or displaying models might end up not working correctly.